Payment Service Added

+++ Hotel.Api.Application/Services/RoomService.cs
@@ -3,6 +3,7 @@ using Hotel.Api.Application.Common.Interfaces;
 using Hotel.Api.Application.Common.Models.Room;
 using Hotel.Api.Domain.Entities;
 using Microsoft.EntityFrameworkCore;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -24,10 +25,21 @@ namespace Hotel.Api.Application.Services
 
         public async Task<int> CreateRoomAsync(RoomCreateModel model)
         {
-            var record = _mapper.Map<Room>(model);
-            await _roomsRepository.InsertAsync(record);
-            model.Images.ForEach(async x => await _roomImageRepository.InsertAsync(new RoomImage { RoomId = record.Id, URL = x }));
-            return record.Id;
+            try
+            {
+                if (await CheckIfRoomNumberExists(model.RoomNumber))
+                {
+                    throw new System.Exception("Room number already exists");
+                }
+                var record = _mapper.Map<Room>(model);
+                await _roomsRepository.InsertAsync(record);
+                model.ImageUrls.ForEach(async x => await _roomImageRepository.InsertAsync(new RoomImage { RoomId = record.Id, URL = x }));
+                return record.Id;
+            }
+            catch
+            {
+                throw;
+            }
         }
 
         public async Task<List<RoomListModel>> GetAllRoomsAsync()
@@ -38,8 +50,19 @@ namespace Hotel.Api.Application.Services
 
         public async Task<RoomModel> GetRoomByIdAsync(int id)
         {
-            var result = await _roomsRepository.GetAsync(query => (IQueryable<Room>)query.Where(x => x.Id == id).Include(x => x.Images).SingleOrDefault());
-            return _mapper.Map<RoomModel>(result);
+            try
+            {
+                var result = await _roomsRepository.GetAsync(query => query.Where(x => x.Id == id).Include(x => x.Images));
+                if (result == null)
+                {
+                    throw new Exception("Room doesn't exist");
+                }
+                return _mapper.Map<RoomModel>(result);
+            }
+            catch (Exception e)
+            {
+                throw;
+            }
         }
 
         public async Task<RoomUpdateModel> UpdateRoomAsync(RoomUpdateModel model)
@@ -49,15 +72,23 @@ namespace Hotel.Api.Application.Services
                 var editeEntity = await _roomsRepository.GetByIdAsync(model.Id);
                 if (editeEntity == null)
                 {
-                    throw new System.Exception("Entity doesn't exist");
+                    throw new Exception("Entity doesn't exist");
                 }
 
                 editeEntity.Capacity = model.Capacity;
                 editeEntity.FloorNumber = model.FloorNumber;
-                editeEntity.RoomNumber = model.RoomNumber;
                 editeEntity.Price = model.Price;
 
                 _roomsRepository.Update(editeEntity);
+
+                foreach (var item in model.ImageURL)
+                {
+                    var check = await _roomImageRepository.GetAsync(query => query.Where(x => x.RoomId == model.Id && x.URL == item));
+                    if (check == null)
+                    {
+                        await _roomImageRepository.InsertAsync(new RoomImage { RoomId = model.Id, URL = item });
+                    }
+                }
             }
             catch
             {
@@ -69,8 +100,22 @@ namespace Hotel.Api.Application.Services
         public async Task<bool> DeleteAsync(int id)
         {
             var deletedEntity = await _roomsRepository.GetByIdAsync(id);
-            await _roomsRepository.DeleteAsync(deletedEntity);
-            return true;
+            if (deletedEntity != null)
+            {
+                await _roomsRepository.DeleteAsync(deletedEntity);
+                return true;
+            }
+            return false;
+        }
+
+        public async Task<bool> CheckIfRoomNumberExists(int roomNumber)
+        {
+            var result = await _roomsRepository.GetAsync(query => query.Where(x => x.RoomNumber == roomNumber));
+            if (result != null)
+            {
+                return true;
+            }
+            return false;
         }
     }
 }
